windowManager.js: created a state machine with the state saved in couchdb; each js file should activate on the page if the state matches the page, and move to the next page (and state) if no other action is required on that page

+++ transferSearch.js
@@ -58,42 +58,51 @@ function setIndexFromValue( elementId, value, valueMapping ) {
 
 // first check for players to add final price
 
-if ( transferSearchPage.exec(window.location.href) ) {
+var stateDb = new PouchDB('http://localhost:5984/hattrick_state');
+stateDb.get('state').then( ( currentState ) => {
 
-  console.error('test transferSearch.js');
+  if ( currentState && currentState.state === 'TRANSFER_CHECK' && transferSearchPage.exec(window.location.href) ) {
 
-  var searchesDb = new PouchDB('http://localhost:5984/hattrick_searches');
+    console.error('test transferSearch.js');
 
-  // get requested searches
-  searchesDb.allDocs({include_docs: true, limit: 1}).then( (searches) => {
-    if ( searches.rows.length > 0 ) {
-      let search = searches.rows[0].doc;
-      console.error('doing search', search);
-      searchesDb.remove( search._id, search._rev );
+    var searchesDb = new PouchDB('http://localhost:5984/hattrick_searches');
 
-      delete search._id;
-      delete search._rev;
+    // get requested searches
+    searchesDb.allDocs({include_docs: true, limit: 1}).then( (searches) => {
+      if ( searches.rows.length > 0 ) {
+        let search = searches.rows[0].doc;
+        console.error('doing search', search);
+        searchesDb.remove( search._id, search._rev );
 
-      setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlAgeMin', search.age.min);
-      setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlAgeMax', search.age.max);
+        delete search._id;
+        delete search._rev;
 
-      let counter = 1;
-      for ( property in search )  {
-        if (property !== 'age') {
-          console.error('setting', property, search[property]);
-          setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter), property);
-          console.error('setting min', property, search[property]);
-          setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter, 'Min'), search[property].min, propertiesValues);
-          console.error('setting max', property, search[property]);
-          setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter, 'Max'), search[property].max, propertiesValues);
+        setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlAgeMin', search.age.min);
+        setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlAgeMax', search.age.max);
+
+        let counter = 1;
+        for ( property in search )  {
+          if (property !== 'age') {
+            console.error('setting', property, search[property]);
+            setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter), property);
+            console.error('setting min', property, search[property]);
+            setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter, 'Min'), search[property].min, propertiesValues);
+            console.error('setting max', property, search[property]);
+            setIndexFromValue('ctl00_ctl00_CPContent_CPMain_ddlSkill'.concat(counter, 'Max'), search[property].max, propertiesValues);
+          }
+          counter = counter + 1;
         }
-        counter = counter + 1;
+        document.getElementById('ctl00_ctl00_CPContent_CPMain_butSearch').click();
+      } else {
+        console.error('no search to perform');
+        currentState.state = 'INACTIVE';
+        stateDb.put(currentState);
       }
-      document.getElementById('ctl00_ctl00_CPContent_CPMain_butSearch').click();
-    } else {
-      console.error('no search to perform');
-    }
 
-  });
+    });
 
-}
+  }
+}).catch( () => {
+  console.warn('state is not set');
+});
+
